# business/tecma-digital-platform/followup-3.0/be-followup-v3/openapi/openapi.v1.yaml
{
  "openapi": "3.0.3",
  "info": {
    "title": "Followup 3.0 REST API",
    "version": "1.0.0",
    "description": "API REST per Followup 3.0 (CRM multi-progetto). Gestione clienti, appartamenti, trattative (rent/sell), calendario, workspaces, progetti, audit e report. Autenticazione JWT (login, refresh). In locale il backend è su porta 8080."
  },
  "servers": [
    {
      "url": "http://localhost:8080/v1",
      "description": "Backend locale (dev)"
    },
    {
      "url": "/v1",
      "description": "Stesso host (proxy)"
    }
  ],
...
  "paths": {
...
    "/apartments/query": {
      "post": {
        "tags": [
          "apartments",
          "Riusabili"
        ],
        "summary": "Query appartamenti paginata (workspaceId, projectIds, search, sort, filters)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "workspaceId",
                  "projectIds"
                ],
                "properties": {
                  "workspaceId": {
                    "type": "string"
                  },
                  "projectIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "page": {
                    "type": "integer",
                    "minimum": 1,
                    "default": 1
                  },
                  "perPage": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 200,
                    "default": 25
                  },
                  "searchText": {
                    "type": "string"
                  },
                  "sort": {
                    "type": "object",
                    "properties": {
                      "field": {
                        "type": "string"
                      },
                      "direction": {
                        "type": "integer",
                        "enum": [
                          1,
                          -1
                        ]
                      }
                    }
                  },
                  "filters": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Lista appartamenti con paginazione",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ApartmentRow"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token mancante o non valido"
          }
        }
      }
    },
...
    "/apartments": {
      "post": {
        "tags": [
          "apartments"
        ],
        "summary": "Create apartment",
        "responses": {
          "200": {
            "description": "Apartment created"
          },
          "409": {
            "description": "Duplicate business conflict"
          }
        }
      }
    },
    "/apartments/{id}": {
      "get": {
        "tags": [
          "apartments"
        ],
        "summary": "Get apartment by id",
        "responses": {
          "200": {
            "description": "Apartment found"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "patch": {
        "tags": [
          "apartments"
        ],
        "summary": "Update apartment by id",
        "responses": {
          "200": {
            "description": "Apartment updated"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token from login or refresh"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Platform consumer API key (scoped by workspace/project)."
      },
      "signatureWebhookAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-signature-webhook-secret",
        "description": "Segreto condiviso per webhook firme (stesso valore di SIGNATURE_WEBHOOK_SECRET). Equivalente: Authorization Bearer <secret>."
      }
    },
    "schemas": {
      "Pagination": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer"
          },
          "perPage": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "totalPages": {
            "type": "integer"
          }
        }
      },
...
      "ApartmentRow": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "workspaceId": {
            "type": "string"
          },
          "projectId": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  }
}